logcat: remove the rest of liblogcat
There were big changes made to support liblogcat, which has long since
been removed.  This change removes the rest of those changes.

It moves the rest of the global state into a class, particularly for
the logcatd case.  The original code re-uses the same context, but
that doesn't seem right or safe.

Test: logcat works, logcat-unit-tests
Test: logpersist works, including last log cat and log rotation
Change-Id: Iee6a2a0319265e87be0bc8fec2b11e8fd2b65ed4

Former-commit-id: d162f14d6cd4b8f1e230164972e5d7d82e3508cf